LayoutTests/platform/ios/imported/w3c/web-platform-tests/selection/fire-selectionchange-event-on-pressing-backspace-expected.txt

world


FAIL Selectionchange event is fired after removing the character assert_equals: target.innerHTML should match expected "<p>wo<i id=\"caret\">l</i>d</p>" but got "<p>wor<i id=\"caret\">l</i>d</p>"
FAIL Selectionchange event is fired after removing the range assert_equals: target.innerHTML should match expected "<p>w<i id=\"caret\">l</i>d</p>" but got "<p>wor<i id=\"caret\">l</i>d</p>"
